Add Docker support
HELLO DUCKER!
…
Ahem. This commit adds some stuff that is useful in case if you want to deploy
Kofta on Docker — including the Dockerfile for Kofta image itself and Docker
Compose config to deploy the whole Kofta stack — that is, Redis, Influx and
Kofta itself (app and one Sidekiq worker) — using one `docker-compose up`
command.

Not tested enough, so bug reports are welcome! Must work on any more or less
modern Linux-based Docker system, and should work on SmartOS as well (to be
tested).